Split Components and Targets
Targets are instrumented chips.
Components are modules that appear in targets.

Targets could possibly be connected to Crobe through more than
one bus (think SWD + Trace + UART).  We will later need a mapping
phase where we explore the discovered components and attach them
to targets.  This is yet to be done.